Sends user to SEND_PAYMENT_SUCCESS_ROUTE on successful reverse submarine swap Renames lightning-result.tsx to lightning-show-invoice.tsx
